multi-utc: put color coding of times on logarithmic scale.
This means for example a system that's off by eight minutes, is one color
further along the scale than a system that's off by four minutes.

Also, because monitor-agent is not meant to communicate with the monitor
second-by-second in real time, I raised the leniency threshhold on times
being considered perfectly in sync (white by default) to 90 seconds, so
that running monitor-agent once a minute on a cron job is good enough.