Standardize my terminal-based widget-handling library.
I've been using essentially the same code in several codebases, and
it's time to unify and generalize it into a library.

The main benefit to server-monitor is improved color handling,
particularly for terminals that don't support 24-bit color, and for
now, this is a limited benefit, since extending the color set in
monitor.cfg is not yet supported.  For backwards compatibility
with existing configurations, specifying colors as rgb triplets
is still supported, but it compromises 8-bit and 4-bit color
support, so it's deprecated.  It'll become *more* deprecated
once user-defined named colors are a thing.