widgets.pl cleanup, and make diffuse a standard widget.
The diffuse widget makes such a nice background, that I decided to
include it in the standard set.

Apart fromn that, this is a huge round of cleanup and bug fixes,
mostly pertaining to background colors, defaults, and ensuring
that everything is always defined that should be.

This commit leaves some croak statements in place that break
backward compatibility with old monitor.cfg files that specify
rgb triplets instead of color names.  I intend to remove those
after further debugging, but for now, be advised that backward
compatibility is temporarily broken, so maybe hold off on
upgrading to this commit if that is an issue for you.